Report errors if CI fails

+++ bin/sbt-ci.sh
@@ -4,9 +4,13 @@ set -o nounset
 # I don't trust $DRONE_COMMIT_MESSAGE here, it's had problems in the future
 if git log -1 --pretty=%B --no-merges | grep DOCS > /dev/null 2>&1; then
   echo "Found DOCS in commit message, the CI will only update the docs site."
-  sbt -sbt-dir /drone/.sbt/1.0 -sbt-boot /drone/.sbt/boot docs/makeSite
+  if ! sbt -sbt-dir /drone/.sbt/1.0 -sbt-boot /drone/.sbt/boot docs/makeSite; then
+    exit 1
+  fi
 else
-  sbt -sbt-dir /drone/.sbt/1.0 -sbt-boot /drone/.sbt/boot $@
+  if ! sbt -sbt-dir /drone/.sbt/1.0 -sbt-boot /drone/.sbt/boot "$@"; then
+    exit 1
+  fi
   find "/root/.ivy2/cache"     -name "ivydata-*.properties" -print -delete
   find "/root/.coursier"       -name "ivydata-*.properties" -print -delete
   find "/root/.sbt"            -name "*.lock"               -print -delete
